exchange response들 수정
