+++ lib/src/models/exchange/exchange_model.dart
@@ -1,31 +1,5 @@
 import 'package:navada_mobile_app/src/models/product/product_model.dart';
-
-class ExchangeModel {
-  bool? success;
-  int? code;
-  String? message;
-  Exchange? data;
-
-  ExchangeModel({this.success, this.code, this.message, this.data});
-
-  ExchangeModel.fromJson(Map<String, dynamic> json) {
-    success = json['success'];
-    code = json['code'];
-    message = json['message'];
-    data = json['data'] != null ? Exchange.fromJson(json['data']) : null;
-  }
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = <String, dynamic>{};
-    data['success'] = success;
-    data['code'] = code;
-    data['message'] = message;
-    if (this.data != null) {
-      data['data'] = this.data!.toJson();
-    }
-    return data;
-  }
-}
+import 'package:navada_mobile_app/src/utilities/enums.dart';
 
 class Exchange {
   String? createdDate;
@@ -35,7 +9,7 @@ class Exchange {
   ProductModel? requesterProduct;
   bool? acceptorConfirmYn;
   bool? requesterConfirmYn;
-  bool? exchangeCompleteYn;
+  ExchangeStatusCd? exchangeStatusCd;
   String? exchangeCompleteDt;
   double? acceptorRating;
   double? requesterRating;
@@ -50,7 +24,7 @@ class Exchange {
       this.requesterProduct,
       this.acceptorConfirmYn,
       this.requesterConfirmYn,
-      this.exchangeCompleteYn,
+      this.exchangeStatusCd,
       this.exchangeCompleteDt,
       this.acceptorRating,
       this.requesterRating,
@@ -69,7 +43,7 @@ class Exchange {
         : null;
     acceptorConfirmYn = json['acceptorConfirmYn'];
     requesterConfirmYn = json['requesterConfirmYn'];
-    exchangeCompleteYn = json['exchangeCompleteYn'];
+    exchangeStatusCd = ExchangeStatusCd.codeToEnum(json['exchangeStatusCd']);
     exchangeCompleteDt = json['exchangeCompleteDt'];
     acceptorRating = json['acceptorRating'];
     requesterRating = json['requesterRating'];
@@ -78,7 +52,7 @@ class Exchange {
   }
 
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> data = Map<String, dynamic>();
+    final Map<String, dynamic> data = <String, dynamic>{};
     data['createdDate'] = createdDate;
     data['modifiedDate'] = modifiedDate;
     data['exchangeId'] = exchangeId;
@@ -90,8 +64,8 @@ class Exchange {
     }
     data['acceptorConfirmYn'] = acceptorConfirmYn;
     data['requesterConfirmYn'] = requesterConfirmYn;
-    data['exchangeCompleteYn'] = exchangeCompleteYn;
     data['exchangeCompleteDt'] = exchangeCompleteDt;
+    data['exchangeStatusCd'] = exchangeStatusCd!.code.toString();
     data['acceptorRating'] = acceptorRating;
     data['requesterRating'] = requesterRating;
     data['acceptorHistoryDeleteYn'] = acceptorHistoryDeleteYn;
